refactor: move config to standalone module and keep debug routes
